vtpm: add method to get sha256 PCRs from Quote
Internally the PCRs are still represented as `Vec<Vec<u8>`, due to
serialization concerns of large arrays. For the consumer a fixed byte
array `Vec<[u8; 32]` is more convenient, so you don't have to perform
runtime assertions on the vector size.

A test has been added to make sure the transformation is
non-destructive.

Signed-off-by: Magnus Kulke <magnuskulke@microsoft.com>